% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/himaconf.R
\name{himaconf}
\alias{himaconf}
\title{High dimensional mediation analysis.}
\usage{
himaconf(
  X,
  M,
  Y,
  COV,
  k = 2,
  CONFmethod = c("RA", "PS", "IPW", "OW"),
  MTmethod = c("HIMA", "HIMA2"),
  tips = TRUE,
  simplify = TRUE
)
}
\arguments{
\item{X}{Binary exposure.}

\item{M}{High dimensional mediators.}

\item{Y}{Continuous outcome.}

\item{COV}{Covariates, also called confounders.}

\item{k}{SIS parameter.}

\item{CONFmethod}{Method of confounding adjustments, including RA,PS,IPW,OW.}

\item{MTmethod}{Method of mediation testing, including HIMA, HIMA2.}

\item{tips}{Whether to output message.}

\item{simplify}{Only output mediator which p value less than 0.05.}
}
\value{
A dataframe contain mediators' point estimate and p value.
}
\description{
High dimensional mediation analysis with confounding adjustment.
Methods of confounding adjustment include regression adjustment,
propensity score adjustment, inverse probability weighting, overlapping weighting.
}
\examples{
library(mydemo)
dat <- sim_data(n = 300, p = 2000)
X <- dat$X
Y <- dat$Y
M <- dat$M
COV <- dat$COV
himaconf(X, M, Y, COV, CONFmethod = "OW", MTmethod = "HIMA2")
}
