% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{sim_data}
\alias{sim_data}
\title{Generate simulated data.}
\usage{
sim_data(
  n,
  p = 2000,
  r = 0.3,
  alpha1.8 = c(0.4, 0.4, 0.5, 0.5, 0.5, 0.5, 0, 0),
  beta1.8 = c(0.4, 0.5, 0.5, 0.6, 0, 0, 0.5, 0.5)
)
}
\arguments{
\item{n}{Sample size.}

\item{p}{Dimension of mediator.}

\item{r}{Covariance between M's error.}

\item{alpha1.8}{Coefficients \code{alpha} of 1st-8th mediator.}

\item{beta1.8}{Coefficients \code{beta} of 1st-8th mediator.}
}
\value{
A list contain X, M, Y, Cov.
}
\description{
Generate simulated data with predefined parameter.
}
\examples{
library(mydemo)
dat <- sim_data(n = 500)
X <- dat$X
Y <- dat$Y
M <- dat$M
COV <- dat$COV
}
